Moved monitor definitions into a seperate file
